Get the main servicex part working
#51
