+++ func_adl_xAOD/__init__.py
@@ -1,3 +1,3 @@
 # Main two things from this package
 # flake8: noqa
-from .use_exe_servicex import FuncADLServerException, use_exe_servicex  # NOQA
+from .ServiceX import ServiceXDatasetSource, FuncADLServerException  # NOQA
