+++ setup.py
@@ -2,7 +2,7 @@
 from setuptools import find_packages  # noqa: F401
 from distutils.core import setup
 from os import listdir
-from version_info import version_servicex
+from version_info import version_func_adl, version_servicex
 
 # Use the readme as the long description.
 with open("README.md", "r") as fh:
@@ -29,7 +29,7 @@ setup(name="func_adl_xAOD",
           "uproot~=3.7",
           "retry~=0.9",
           "jinja2",
-          "func_adl.ast==1.1.0b1",
+          f"func_adl.ast=={version_func_adl}",
           "qastle==0.7",
           f"servicex=={version_servicex}"
       ],
